adapt to supplier refactoring in jicoco
