+++ jvb/src/main/java/org/jitsi/videobridge/xmpp/XmppCommon.java
@@ -203,7 +203,7 @@ public class XmppCommon
      */
     private IQ handleVersionIQ(Version versionRequest)
     {
-        VersionService versionService = VersionServiceSupplierKt.singleton.get();
+        VersionService versionService = JvbVersionServiceSupplierKt.singleton.get();
 
         org.jitsi.utils.version.Version currentVersion
                 = versionService.getCurrentVersion();
