+++ jvb/src/main/kotlin/org/jitsi/videobridge/rest/binders/ServiceBinder.kt
@@ -17,8 +17,12 @@
 package org.jitsi.videobridge.rest.binders
 
 import org.glassfish.hk2.utilities.binding.AbstractBinder
+import org.jitsi.health.HealthCheckServiceSupplier
+import org.jitsi.version.VersionServiceSupplier
 import org.jitsi.videobridge.VideobridgeSupplier
+import org.jitsi.videobridge.health.singleton as healthCheckServiceSupplierSingleton
 import org.jitsi.videobridge.stats.StatsManagerSupplier
+import org.jitsi.videobridge.version.singleton
 import org.jitsi.videobridge.xmpp.ClientConnectionImplSupplier
 import org.jitsi.videobridge.singleton as videobridgeSupplierSingleton
 import org.jitsi.videobridge.stats.singleton as statsManagerSupplierSingleton
@@ -29,5 +33,7 @@ class ServiceBinder : AbstractBinder() {
         bind(videobridgeSupplierSingleton).to(VideobridgeSupplier::class.java)
         bind(statsManagerSupplierSingleton).to(StatsManagerSupplier::class.java)
         bind(clientConnectionImplSupplierSingleton).to(ClientConnectionImplSupplier::class.java)
+        bind(healthCheckServiceSupplierSingleton).to(HealthCheckServiceSupplier::class.java)
+        bind(singleton).to(VersionServiceSupplier::class.java)
     }
 }
